feat: Implement automatic model selection across all 3 projects
- Add auto-selection logic to training pipelines in BankChurn, CarVision, and TelecomAI
- Training now compares all configured models and automatically selects the best one based on metrics
- BankChurn: Auto-selected RandomForest (F1=0.8453) over Ensemble (F1=0.8160) - +3.6% improvement
- CarVision: Auto-selected XGBoost (R²=0.7048) over RandomForest (R²=0.6753) - +4.4% improvement
- TelecomAI: Kept XGBoost (F1=0.8086) as it was already the best model
- Fix random_state duplication bug in neural_network factory functions
- Fix model_year dtype issue in CarVision (int16 -> float32 to handle NaN)
- Extend CarVision config schema to support all model types (xgboost, lightgbm, neural_network)
- Update all 3 model.joblib files with the best performing models
- All tests passing (BankChurn: ✓, CarVision: ✓, TelecomAI: 50/50 unit tests ✓)

+++ CarVision-Market-Intelligence/src/carvision/data.py
@@ -31,7 +31,7 @@ def load_data(csv_path: str) -> pd.DataFrame:
     # Note: cylinders is not specified as it may contain strings like '6 cylinders'
     dtypes = {
         "price": "float32",
-        "model_year": "int16",
+        "model_year": "float32",
         "odometer": "float32",
         "condition": "category",
         "fuel": "category",
